pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vllm-wizard"
version = "0.1.0"
description = "CLI tool for vLLM configuration generation and GPU sizing"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.9"
authors = [
    { name = "Vadim Vashkelis", email = "vashkelis@gmail.com" }
]
keywords = ["vllm", "llm", "gpu", "sizing", "configuration"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.0",
    "typer>=0.12",
    "rich>=13.0",
    "pyyaml>=6.0",
    "huggingface_hub>=0.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.3",
]
web = [
    "fastapi>=0.110",
    "uvicorn>=0.28",
    "jinja2>=3.1",
]

[project.scripts]
vllm-wizard = "vllm_wizard.cli:app"

[project.urls]
Homepage = "https://github.com/vashkelis/vllm-wizard"
Documentation = "https://github.com/vashkelis/vllm-wizard#readme"
Repository = "https://github.com/vashkelis/vllm-wizard"

[tool.hatch.build.targets.wheel]
packages = ["src/vllm_wizard"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
